+++ src/main/kotlin/com/github/blarc/gitlab/template/lint/plugin/gitlab/GitlabFactory.kt
@@ -1,17 +1,17 @@
 package com.github.blarc.gitlab.template.lint.plugin.gitlab
 
-import com.github.blarc.gitlab.template.lint.plugin.settings.AppSettings
 import com.intellij.openapi.project.Project
+import java.net.URI
 
 class GitlabFactory() {
 
-    fun getGitLab(url: String): Gitlab {
-        return Gitlab(url, AppSettings.instance?.gitlabToken!!)
+    fun getGitLab(url: URI, gitlabToken: String): Gitlab {
+        return Gitlab("${url.scheme}://${url.host}", gitlabToken)
     }
 
     companion object {
-        fun getInstance(project: Project?): GitlabFactory? {
-            return project!!.getService(GitlabFactory::class.java)
+        fun getInstance(project: Project): GitlabFactory? {
+            return project.getService(GitlabFactory::class.java)
         }
     }
 }