+++ src/main/kotlin/com/github/blarc/gitlab/template/lint/plugin/inspections/GitlabLintInspection.kt
@@ -2,7 +2,6 @@ package com.github.blarc.gitlab.template.lint.plugin.inspections
 
 import com.github.blarc.gitlab.template.lint.plugin.GitlabLintUtils.Companion.matchesGitlabLintRegex
 import com.github.blarc.gitlab.template.lint.plugin.runLinting
-import com.github.blarc.gitlab.template.lint.plugin.settings.AppSettings
 import com.intellij.codeInspection.LocalInspectionTool
 import com.intellij.codeInspection.ProblemsHolder
 import com.intellij.openapi.components.Service
@@ -21,10 +20,7 @@ class GitlabLintInspection : LocalInspectionTool() {
                     indicator.isIndeterminate = true
                     val project = file.project
                     if (matchesGitlabLintRegex(file.name)) {
-                        val gitlabToken = AppSettings.instance?.gitlabToken
-                        if (!gitlabToken.isNullOrEmpty()) {
-                            runLinting(project, file)
-                        }
+                        runLinting(project, file)
                     }
                     EditorNotifications.getInstance(project).updateAllNotifications()
                 }