+++ src/main/kotlin/com/github/blarc/gitlab/template/lint/plugin/pipeline/middleware/RecordHit.kt
@@ -3,6 +3,7 @@ package com.github.blarc.gitlab.template.lint.plugin.pipeline.middleware
 import com.github.blarc.gitlab.template.lint.plugin.gitlab.GitlabLintResponse
 import com.github.blarc.gitlab.template.lint.plugin.pipeline.Pass
 import com.github.blarc.gitlab.template.lint.plugin.settings.AppSettings
+import com.github.blarc.gitlab.template.lint.plugin.widgets.LintStatusEnum
 import com.intellij.openapi.components.Service
 import com.intellij.openapi.components.service
 
@@ -10,12 +11,12 @@ import com.intellij.openapi.components.service
 class RecordHit : Middleware {
     override val priority = 20
 
-    override fun invoke(pass: Pass, next: () -> GitlabLintResponse?) : GitlabLintResponse? {
-        val gitlabLintResponse = next() ?: return null
+    override fun invoke(pass: Pass, next: () -> Pair<GitlabLintResponse?, LintStatusEnum>?) : Pair<GitlabLintResponse?, LintStatusEnum>? {
+        val result = next() ?: return null
 
         service<AppSettings>().recordHit();
 
-        return gitlabLintResponse;
+        return result;
     }
 }