+++ src/main/kotlin/com/github/blarc/gitlab/template/lint/plugin/pipeline/middleware/ResolveContext.kt
@@ -1,13 +1,15 @@
 package com.github.blarc.gitlab.template.lint.plugin.pipeline.middleware
 
-import com.github.blarc.gitlab.template.lint.plugin.gitlab.GitlabLintResponse
 import com.github.blarc.gitlab.template.lint.plugin.git.httpUrl
 import com.github.blarc.gitlab.template.lint.plugin.git.locateRemote
 import com.github.blarc.gitlab.template.lint.plugin.gitlab.GitlabFactory
+import com.github.blarc.gitlab.template.lint.plugin.gitlab.GitlabLintResponse
 import com.github.blarc.gitlab.template.lint.plugin.notifications.Notification
 import com.github.blarc.gitlab.template.lint.plugin.notifications.sendNotification
 import com.github.blarc.gitlab.template.lint.plugin.pipeline.Pass
+import com.github.blarc.gitlab.template.lint.plugin.settings.AppSettings
 import com.github.blarc.gitlab.template.lint.plugin.settings.ProjectSettings
+import com.github.blarc.gitlab.template.lint.plugin.widgets.LintStatusEnum
 import com.intellij.openapi.components.Service
 import com.intellij.openapi.components.service
 import git4idea.repo.GitRemote
@@ -17,12 +19,15 @@ import git4idea.repo.GitRepositoryManager
 @Service
 class ResolveContext : Middleware {
     override val priority = 5
+    private var showGitlabTokenNotification = true
 
-    override fun invoke(pass: Pass, next: () -> GitlabLintResponse?): GitlabLintResponse? {
+    override fun invoke(pass: Pass, next: () -> Pair<GitlabLintResponse?, LintStatusEnum>?): Pair<GitlabLintResponse?, LintStatusEnum>? {
         val repository = locateRepository(pass) ?: return null
         val remote = locateRemote(pass, repository) ?: return null
+        val url = remote.httpUrl ?: return null
+        val gitlabToken = getGitlabToken(pass) ?: return null
 
-        pass.gitlab = GitlabFactory.getInstance(pass.project)!!.getGitLab("${remote.httpUrl?.scheme}://${remote.httpUrl?.host}")
+        pass.gitlab = GitlabFactory.getInstance(pass.project)!!.getGitLab(url, gitlabToken)
         pass.repository = repository
         pass.remote = remote
 
@@ -43,5 +48,19 @@ class ResolveContext : Middleware {
         remote ?: sendNotification(Notification.remoteNotFound(), pass.project)
 
         return remote
+    }
+
+    private fun getGitlabToken(pass: Pass) : String? {
+        val gitlabToken = AppSettings.instance?.gitlabToken
+
+        if (gitlabToken.isNullOrEmpty() && showGitlabTokenNotification) {
+            sendNotification(Notification.gitlabTokenNotSet(pass.project), pass.project)
+            showGitlabTokenNotification = false
+        }
+        else if (!gitlabToken.isNullOrEmpty()) {
+            showGitlabTokenNotification = true
+        }
+
+        return gitlabToken
     }
 }