+++ src/main/kotlin/com/github/blarc/gitlab/template/lint/plugin/pipeline/middleware/ShowSupportNotification.kt
@@ -5,6 +5,7 @@ import com.github.blarc.gitlab.template.lint.plugin.notifications.Notification
 import com.github.blarc.gitlab.template.lint.plugin.notifications.sendNotification
 import com.github.blarc.gitlab.template.lint.plugin.pipeline.Pass
 import com.github.blarc.gitlab.template.lint.plugin.settings.AppSettings
+import com.github.blarc.gitlab.template.lint.plugin.widgets.LintStatusEnum
 import com.intellij.openapi.components.Service
 import com.intellij.openapi.components.service
 
@@ -12,7 +13,7 @@ import com.intellij.openapi.components.service
 class ShowSupportNotification : Middleware {
     override val priority = 10
 
-    override fun invoke(pass: Pass, next: () -> GitlabLintResponse?): GitlabLintResponse? {
+    override fun invoke(pass: Pass, next: () -> Pair<GitlabLintResponse?, LintStatusEnum>?): Pair<GitlabLintResponse?, LintStatusEnum>? {
         val gitlabLintResponse = next()
 
         val settings = service<AppSettings>()